pyproject.toml
[project]
name = "nbachat_backend"
version = "0.1.0"
description = ""
authors = [
    {name = "CommonGibbon",email = "xenoproboscizoid@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = ["nba-api (>=1.10.2,<2.0.0)", "langchain (>=1.0.3,<2.0.0)", "mcp (>=1.20.0,<2.0.0)", "pandas (>=2.3.3,<3.0.0)", "python-toon (>=0.1.2,<0.2.0)", "openai (>=2.6.1,<3.0.0)", "fastapi (>=0.120.4,<0.121.0)", "uvicorn (>=0.38.0,<0.39.0)", "psycopg2-binary (>=2.9.11,<3.0.0)", "sqlalchemy (>=2.0.44,<3.0.0)", "python-dotenv (>=1.2.1,<2.0.0)", "tabulate (>=0.9.0,<0.10.0)", "websockets (>=15.0.1,<16.0.0)", "google-adk (>=1.18.0,<2.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.10"
ipykernel = "^7.1.0"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.poetry]
packages = [
  { include = "server", from = "src" },
  { include = "nba_mcp_server", from = "src" }
]